Refactor Observation Functions into Shared Clinical Module
Refactors `create_clinical_result_observation_resource`, `mapped_snomed_test_results`,
and `present` into the shared clinical module. Combines the functionality from the
`virology_test_result` function into the refactored `mapped_snomed_test_results`
function. Ensures that we only ingest inconclusive results for sites that desire them.
Also changes the SNOMED code for Covid testing from the International version to the
US version.
