+++ lib/seattleflu/id3c/cli/command/etl/redcap_det_uw_retrospectives.py
@@ -443,156 +443,6 @@ def determine_questionnaire_items(record: dict) -> List[dict]:
     return questionnaire_items
 
 
-def create_clinical_result_observation_resource(redcap_record: dict) -> Optional[List[dict]]:
-    """
-    Determine the clinical results based on responses in *redcap_record* and
-    create observation resources for each result following the FHIR format
-    (http://www.hl7.org/implement/standards/fhir/observation.html)
-    """
-    code_map = {
-        '1240581000000104': {
-            'system': 'http://snomed.info/sct',
-            'code': '1240581000000104',
-            'display': '2019-nCoV (novel coronavirus) detected',
-        },
-        '181000124108': {
-            'system': 'http://snomed.info/sct',
-            'code': '181000124108',
-            'display': 'Influenza A virus present',
-        },
-        '441345003': {
-            'system': 'http://snomed.info/sct',
-            'code': '441345003',
-            'display': 'Influenza B virus present',
-        },
-        '441278007': {
-            'system': 'http://snomed.info/sct',
-            'code': '441278007',
-            'display': 'Respiratory syncytial virus untyped strain present',
-        },
-        '440925005': {
-            'system': 'http://snomed.info/sct',
-            'code': '440925005',
-            'display': 'Human rhinovirus present',
-        },
-        '440930009': {
-            'system': 'http://snomed.info/sct',
-            'code': '440930009',
-            'display': 'Human adenovirus present',
-        },
-    }
-
-    # Create intermediary, mapped clinical results using SNOMED codes.
-    # This is useful in removing duplicate tests (e.g. multiple tests run for
-    # Influenza A)
-    results = mapped_snomed_test_results(redcap_record)
-
-    # Create observation resources for all results in clinical tests
-    diagnostic_results: Any = {}
-
-    for index, finding in enumerate(results):
-        new_observation = observation_resource('clinical')
-        new_observation['id'] = 'result-' + str(index + 1)
-        new_observation['code']['coding'] = [code_map[finding]]
-        new_observation['valueBoolean'] = results[finding]
-        diagnostic_results[finding] = (new_observation)
-
-    return list(diagnostic_results.values()) or None
-
-
-def present(redcap_record: dict, test: str) -> Optional[bool]:
-    """
-    Returns a mapped boolean *test* result from a given *redcap_record*. A
-    return value of `None` means no test results are available.
-    """
-    result = redcap_record[test]
-
-    # Lowercase, remove non-alpahnumeric characters(except spaces), then standardize whitespace
-    # Removal of non-alphanumeric characters is to account for inconsistencies in the data received
-    standardized_result = standardize_whitespace(re.sub(r'[^a-z0-9 ]+','',result.lower())) if result else None
-
-    if not standardized_result:
-        return None
-
-    test_result_prefix_map = {
-        'negative'                              : False,
-        'none detected'                         : False,
-        'not detected'                          : False,
-        'detected'                              : True,
-        'positive'                              : True,
-        'cancel'                                : None,
-        'disregard'                             : None,
-        'duplicate request'                     : None,
-        'inconclusive'                          : None, # XXX: Ingest this someday as present = null?
-        'indeterminate'                         : None, # XXX: Ingest this someday as present = null?
-        'pending'                               : None,
-        'test not applicable'                   : None,
-        'wrong test'                            : None,
-        'followup testing required'             : None,
-        'data entry correction'                 : None,
-        'reorder requested'                     : None,
-        'invalid'                               : None,
-    }
-
-    for prefix in test_result_prefix_map:
-        if standardized_result.startswith(prefix):
-            return test_result_prefix_map[prefix]
-
-    raise UnknownTestResult(f"Unknown test result value «{standardized_result}» for «{redcap_record['barcode']}».")
-
-
-def mapped_snomed_test_results(redcap_record: dict) -> Dict[str, bool]:
-    """
-    Given a *redcap_record*, returns a dict of the mapped SNOMED clinical
-    finding code and the test result.
-    """
-    # I'm using a British version (1) of snomed for COVID-19 rather than the
-    # international verison, because it appears there is still no
-    # observation result for SARS-CoV-2 in the latest international edition
-    # (2).
-    #
-    # 1. https://snomedbrowser.com/Codes/Details/1240581000000104
-    # 2: https://www.snomed.org/news-and-events/articles/march-2020-interim-snomedct-release-COVID-19
-    #
-    # -- kfay, 11 Mar 2020
-    redcap_to_snomed_map = {
-        'ncvrt': '1240581000000104',
-        'revfla': '181000124108',
-        'fluapr': '181000124108',
-        'revflb': '441345003',
-        'flubpr': '441345003',
-        'revrsv': '441278007',
-        'revrhn': '440925005',
-        'revadv': '440930009',
-    }
-
-    results: Dict[str, bool] = {}
-
-    # Populate dict of tests administered during encounter by filtering out
-    # null results. Map the REDCap test variable to the snomed code. In the
-    # event of duplicate clinical findings, prioritize keeping positive results.
-    for test in redcap_to_snomed_map:
-        code = redcap_to_snomed_map[test]
-
-        # Skip updating results for tests already marked as positive
-        if results.get(code):
-            continue
-
-        try:
-            result = present(redcap_record, test)
-        except UnknownTestResult as e:
-            LOG.warning(e)
-            continue
-
-        # Don't add empty or inconclusive results
-        if result is None:
-            continue
-
-        results[code] = result
-
-    return results
-
-
 class UnknownHospitalDischargeDisposition(ValueError):
     """
     Raised by :function: `discharge_disposition` if it finds
